Add function remoteEvalOnWorkerWithCustomLog
Allows to specify custom log path for a new worker.
This could be useful e.g. when we have to start many workers, but do want to create too many log files.
In that case, we can have a pool of log files (one log per CPU) and reuse them.